requirements.txt
jeepney
requests
opencv-python
dbus-python==1.2.18
pygobject